+++ src/syft_rds/client/local_stores/base.py
@@ -2,12 +2,12 @@ from typing import TYPE_CHECKING, ClassVar, Generic, List, Type, TypeVar
 
 from syft_core import Client as SyftBoxClient
 
-from syft_rds.jupyter_utils.types import TableList
-from syft_rds.models.base import ItemBase, ItemBaseCreate, ItemBaseUpdate
-from syft_rds.models.models import (
+from syft_rds.display_utils.jupyter.types import TableList
+from syft_rds.models import (
     GetAllRequest,
     GetOneRequest,
 )
+from syft_rds.models.base import ItemBase, ItemBaseCreate, ItemBaseUpdate
 from syft_rds.store import YAMLStore
 
 if TYPE_CHECKING:
