+++ src/syft_rds/client/local_stores/job.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Final, Type
 
 from syft_rds.client.local_stores.base import CRUDLocalStore
-from syft_rds.models.models import Job, JobCreate, JobUpdate
+from syft_rds.models import Job, JobCreate, JobUpdate
 
 
 class JobLocalStore(CRUDLocalStore[Job, JobCreate, JobUpdate]):
