+++ src/syft_rds/client/local_stores/runtime.py
@@ -1,7 +1,7 @@
 from typing import Final, Type
 
 from syft_rds.client.local_stores.base import CRUDLocalStore
-from syft_rds.models.models import Runtime, RuntimeCreate, RuntimeUpdate
+from syft_rds.models import Runtime, RuntimeCreate, RuntimeUpdate
 
 
 class RuntimeLocalStore(CRUDLocalStore[Runtime, RuntimeCreate, RuntimeUpdate]):
