+++ src/syft_rds/client/local_stores/user_code.py
@@ -1,7 +1,7 @@
 from typing import Final, Type
 
 from syft_rds.client.local_stores.base import CRUDLocalStore
-from syft_rds.models.models import UserCode, UserCodeCreate, UserCodeUpdate
+from syft_rds.models import UserCode, UserCodeCreate, UserCodeUpdate
 
 
 class UserCodeLocalStore(CRUDLocalStore[UserCode, UserCodeCreate, UserCodeUpdate]):
