+++ src/syft_rds/client/rds_clients/base.py
@@ -7,7 +7,7 @@ from syft_core import Client as SyftBoxClient
 
 from syft_rds.client.local_store import LocalStore
 from syft_rds.client.rpc import RPCClient, T
-from syft_rds.models.models import GetAllRequest, GetOneRequest, Job
+from syft_rds.models import GetAllRequest, GetOneRequest, Job
 from syft_rds.syft_runtime.main import CodeRuntime
 
 if TYPE_CHECKING:
