+++ src/syft_rds/client/rds_clients/runtime.py
@@ -1,5 +1,5 @@
 from syft_rds.client.rds_clients.base import RDSClientModule
-from syft_rds.models.models import Runtime
+from syft_rds.models import Runtime
 
 
 class RuntimeRDSClient(RDSClientModule[Runtime]):
