+++ src/syft_rds/client/rds_clients/user_code.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from syft_rds.client.rds_clients.base import RDSClientModule
 from syft_rds.client.utils import PathLike
-from syft_rds.models.models import (
+from syft_rds.models import (
     UserCode,
     UserCodeCreate,
     UserCodeType,
