+++ src/syft_rds/display_utils/jupyter/types.py
@@ -1,6 +1,7 @@
-from collections import UserDict, UserList
 import os
-from syft_rds.jupyter_utils.tabulator import build_tabulator_table
+from collections import UserDict, UserList
+
+from syft_rds.display_utils.jupyter.tabulator import build_tabulator_table
 
 SYFT_NO_REPR_HTML = "SYFT_NO_REPR_HTML" in os.environ
