+++ src/syft_rds/server/routers/runtime_router.py
@@ -1,5 +1,6 @@
 from syft_event import SyftEvents
-from syft_rds.models.models import (
+
+from syft_rds.models import (
     GetAllRequest,
     GetOneRequest,
     ItemList,
